pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chatbert"
version = "0.1.0"
description = "ChatBERT: Conversational AI with Bidirectional Encoder Representations"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "ChatBERT Authors"}
]
keywords = ["nlp", "bert", "chatbot", "transformers", "deep-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "datasets>=2.16.0",
    "accelerate>=0.25.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
train = [
    "wandb>=0.16.0",
    "tensorboard>=2.15.0",
    "bitsandbytes>=0.41.0",
]
eval = [
    "evaluate>=0.4.0",
    "rouge-score>=0.1.2",
    "bert-score>=0.3.13",
    "nltk>=3.8.0",
]
demo = [
    "gradio>=4.0.0",
]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]
all = ["chatbert[train,eval,demo,dev]"]

[project.urls]
Homepage = "https://github.com/chatbert/chatbert"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100
